+++ src/lib/tabs/tabWarp/WarpSettingsPanel.tsx
@@ -3,7 +3,8 @@ import { Flex, NumberInput, SegmentedControl, Select, Title as MantineTitle } fr
 import type { UseFormReturnType } from '@mantine/form'
 import i18next from 'i18next'
 import { Assets } from 'lib/rendering/assets'
-import { StarlightMultiplier, StarlightRefund, type WarpRequest, WarpIncomeOptions, WarpIncomeType } from 'lib/tabs/tabWarp/warpCalculatorTypes'
+import { WarpIncomeOptions } from 'lib/tabs/tabWarp/warpCalculatorController'
+import { StarlightMultiplier, StarlightRefund, type WarpRequest, WarpIncomeType } from 'lib/tabs/tabWarp/warpCalculatorTypes'
 import { PassIcon } from 'lib/tabs/tabWarp/WarpMilestoneTable'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { MultiSelectPills } from 'lib/ui/MultiSelectPills'
