+++ src/lib/tabs/tabWarp/warpCalculatorController.test.ts
@@ -1,17 +1,15 @@
 // @vitest-environment jsdom
 import { Metadata } from 'lib/state/metadataInitializer'
-import { calculateWarps } from 'lib/tabs/tabWarp/warpCalculatorController'
+import { calculateWarps, normalizeWarpRequest, WarpIncomeOptions } from 'lib/tabs/tabWarp/warpCalculatorController'
 import {
   DEFAULT_WARP_REQUEST,
   DEFAULT_WARP_TARGET,
   EidolonLevel,
   NONE_WARP_INCOME_OPTION,
-  normalizeWarpRequest,
   StarlightRefund,
   SuperimpositionLevel,
   type WarpRequest,
   type WarpTarget,
-  WarpIncomeOptions,
   WarpStrategy,
 } from 'lib/tabs/tabWarp/warpCalculatorTypes'
 import {
